Add an adapter for easily attaching code to specific routes
Before this, routing for apps and blueprints was nested inside the object.
The advantage was that this code could access the configuration values.

Therefore. this changes the blueprint code so that request.bp contains
the current blueprint (i.e. the current SiteBlueprint object) which
allows the code to access configuraton values, permissions, etc.